% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBCSutils-package.R
\docType{package}
\name{BBCSutils-package}
\alias{BBCSutils}
\alias{BBCSutils-package}
\title{BBCSutils: Utilities and general helper R package with various functions}
\description{
This package contains numerous helper functions. There include functions for confidence intervals, predicting SQL data types, ggplot themes and palettes, ploty helpers, binomial thinning, standardisation methods and dispersion metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://birmingham-and-solihull-ics.github.io/BBCSutils/}
  \item Report bugs at \url{https://github.com/birmingham-and-solihull-ics/BBCSutils/issues,}
}

}
\author{
\strong{Maintainer}: Chris Mainey \email{c.mainey1@nhs.net} (\href{https://orcid.org/0000-0002-3018-6171}{ORCID})

Authors:
\itemize{
  \item Chris Mainey \email{c.mainey1@nhs.net} (\href{https://orcid.org/0000-0002-3018-6171}{ORCID})
  \item Siti Hassan \email{siti.hassan@nhs.net} [contributor]
  \item Archie Atack \email{a.atack@nhs.net} [contributor]
  \item Nathan Thomas \email{nathan.thomas6@nhs.net} (\href{https://orcid.org/0009-0006-3734-1943}{ORCID}) [contributor]
}

}
\keyword{internal}
