% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{icb_theme_gradient_colour_scales}
\alias{icb_theme_gradient_colour_scales}
\alias{scale_colour_icb}
\alias{scale_color_icb}
\alias{scale_fill_icb}
\title{ICB Gradient colour Scale}
\usage{
scale_colour_icb(
  palette = c("cluster", "bsol", "alternative", "ics_blue", "ics_orange", "ics_green",
    "ics_purple", "ics_navy", "contrast_green_orange", "contrast_orange_blue",
    "contrast_blue_purple", "contrast_purple_green"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_color_icb(
  palette = c("cluster", "bsol", "alternative", "ics_blue", "ics_orange", "ics_green",
    "ics_purple", "ics_navy", "contrast_green_orange", "contrast_orange_blue",
    "contrast_blue_purple", "contrast_purple_green"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_icb(
  palette = c("cluster", "bsol", "alternative", "ics_blue", "ics_orange", "ics_green",
    "ics_purple", "ics_navy", "contrast_green_orange", "contrast_orange_blue",
    "contrast_blue_purple", "contrast_purple_green"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use}

\item{discrete}{Optional: boolean to indicate that this scale is a discrete
scale. Defaults to TRUE}

\item{reverse}{Optional: boolean to reverse the direction of the scale.
Defaults to FALSE}

\item{...}{additional arguments passed to the ggplot functions}
}
\value{
a scale object to be used with a ggplot object
}
\description{
Generates a colour gradient
}
