% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{icb_theme_pal}
\alias{icb_theme_pal}
\title{ICB Theme Palettes}
\usage{
icb_theme_pal(
  palette = c("cluster", "bsol", "alternative", "ics_blue", "ics_orange", "ics_green",
    "ics_purple", "ics_navy", "contrast_green_orange", "contrast_orange_blue",
    "contrast_blue_purple", "contrast_purple_green"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use. 'cluster' is the BSOL / BC cluster colours and default.}

\item{reverse}{Optional: boolean value to reverse the palette}

\item{...}{arguments to pass to colorRamp}
}
\value{
A colorRamp function
}
\description{
This function returns a palette generated by colorRampPalette from one of the
ICB palettes
}
\examples{
icb_theme_pal("cluster")
icb_theme_pal("bsol")
# or, specify a palette
icb_theme_pal("ics_orange")
# a reverse palette
icb_theme_pal("cluster", TRUE)
}
