% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{icb_theme_pal_vec}
\alias{icb_theme_pal_vec}
\title{Returns ICB theme palette as a vector}
\usage{
icb_theme_pal_vec(palette)
}
\arguments{
\item{palette}{The name of the palette to be used}
}
\value{
A vector of colours from the palette
}
\description{
Returns an ICB theme palette as a vector of colours for use with
scale_fill_manual() and scale_colour_manual(). This allows the exact colours
to be used for a discrete scale rather than colours selected from a gradient
defined by the palette.
}
