% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRs.R
\docType{data}
\name{ISR_example}
\alias{ISR_example}
\title{ISR example dataset}
\format{
\subsection{\code{ISR_example}}{

A data frame with 192 rows and 5 columns:
\describe{
\item{age_group_code}{Numerically coded age-bands}
\item{sex_group_code}{Numerically coded sex groups}
\item{imd_code}{Index of multiple deprivation qunitiles, 1-5, 999 = unknown}
\item{numerator}{Numerator of a given indicator}
\item{denominator}{Denominator of a given indicator}
}
}
}
\source{
Constructed from code \url{https://github.com/BBCS-PHI/BSOLutils/blob/main/data-raw/build_example_data.R}
}
\usage{
data(ISR_example)
}
\description{
An artificial dataset with 6 age bads, 2 sex bands and 5 IMD quintiles plus
and unknown coded as 999.
This is here to test and demonstrate the function
}
\examples{
data(ISR_example)

}
\keyword{datasets}
